Add `--output` parameter and enable JSON printing
This adds a new parameter to determine the type of output that this
script will use. JSON output is now an option, which will take the
previously unused `file` argument and write a json-type output there.
This will be used by the UI.

If no file is given as an output, it'll just print it to stdout.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>